@@ -1,53 +1,52 @@
 {
   "name": "tads",
-  "version": "0.10.0",
-  "private": false,
+  "version": "v3.0.0--beta",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "core-js": "^3.3.2",
-    "minireset.css": "^0.0.6",
-    "v-tooltip": "2.1.3",
-    "vue": "^2.6.10"
+    "core-js": "3.6.5",
+    "lodash": "4.17.21",
+    "minireset.css": "0.0.7",
+    "vue": "3.2.20"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.4",
-    "@storybook/addon-actions": "^5.2.5",
-    "@storybook/addon-backgrounds": "^5.2.5",
-    "@storybook/addon-docs": "^5.2.5",
-    "@storybook/addon-knobs": "^5.2.5",
-    "@storybook/addon-links": "^5.2.5",
-    "@storybook/addon-storysource": "^5.2.5",
-    "@storybook/addons": "^5.2.5",
-    "@storybook/preset-scss": "^1.0.2",
-    "@storybook/source-loader": "^5.2.5",
-    "@storybook/theming": "^6.3.1",
-    "@storybook/vue": "^5.2.5",
-    "@vue/cli-plugin-babel": "^4.0.0",
-    "@vue/cli-plugin-eslint": "^4.0.0",
-    "@vue/cli-plugin-unit-jest": "^4.0.0",
-    "@vue/cli-service": "^4.0.0",
-    "@vue/eslint-config-prettier": "^5.0.0",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "babel-preset-vue": "^2.0.2",
-    "css-loader": "^3.2.0",
-    "eslint": "^5.16.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-vue": "^5.0.0",
-    "node-sass": "^6.0.1",
-    "sass-loader": "^10",
-    "prettier": "^1.18.2",
-    "resolve-url-loader": "^3.1.0",
-    "style-loader": "^1.0.0",
-    "vue-svg-loader": "^0.12.0",
-    "vue-template-compiler": "^2.6.10"
-  }
+    "@babel/core": "^7.15.8",
+    "@storybook/addon-actions": "^6.3.12",
+    "@storybook/addon-essentials": "^6.3.12",
+    "@storybook/addon-links": "^6.3.12",
+    "@storybook/vue3": "^6.3.12",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.2",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^7.0.0",
+    "vue-loader": "^16.8.1"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "babel-eslint"
+    },
+    "rules": {}
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead"
+  ]
 }
