@@ -1,12 +1,5 @@
 module.exports = {
-  presets: ["@vue/app"],
-  plugins: [
-    "@babel/plugin-proposal-object-rest-spread",
-    "@babel/transform-runtime"
-  ],
-  env: {
-    test: {
-      plugins: ["require-context-hook"]
-    }
-  }
-};
+  presets: [
+    '@vue/cli-plugin-babel/preset'
+  ]
+}
