@@ -7,12 +7,10 @@
 
       <select
         ref="select"
-        v-model="computedValue"
+        :dir="dir"
+        v-model="localValue"
         v-bind="$attrs"
-        @blur="$emit('blur', $event)"
-        @focus="$emit('focus', $event)"
         :style="widthStyle"
-        :required="required"
       >
         <template v-if="placeholder">
           <option
@@ -30,8 +28,8 @@
     </span>
 
     <!-- This is needed to measure the selected options text length to adjust the real select's width -->
-    <div class="dummy-wrap">
-      <span class="select dummy" :class="fieldClasses" ref="dummySelect">
+    <div class="select__dummy-wrap">
+      <span class="select select__dummy-wrap__dummy" :class="fieldClasses" ref="dummySelect">
       <select>
         <option>{{ dummyText }}</option>
       </select>
@@ -41,20 +39,23 @@
 </template>
 
 <script>
+import './select.css'
+import {computed} from "vue";
+
 export default {
   name: "TadsSelect",
   inheritAttrs: false,
   props: {
-    value: {
+    modelValue: {
       type: [String, Number, Boolean, Object, Array, Function],
       default: null
     },
     placeholder: {
-      type: [String, Number, Boolean],
+      type: String,
       default: ""
     },
     label: {
-      type: [String, Number, Boolean],
+      type: String,
       default: ""
     },
     small: Boolean,
@@ -62,29 +63,36 @@ export default {
     transparent: Boolean,
     expanded: Boolean,
     autoWidth: Boolean,
-    maxWidth: Number,
-    required: Boolean
+    autoWidthMax: Number,
+    required: Boolean,
+    border: Boolean,
   },
+  emits: ['update:modelValue'],
+  setup(props, context) {
+    const localValue = computed({
+      get: () => props.modelValue,
+      set: (value) => {
+        context.emit('update:modelValue', value)
+        this.calculateWidth()
+      }
+    });
+
+    return {
+      localValue
+    };
+  },
+
   data() {
     return {
       selected: this.value,
       width: false,
-      dummyText: null
+      dummyText: null,
+      dir: (this.label && this.transparent) ? 'rtl' : 'ltf'
     };
   },
 
   mounted() {
     this.calculateWidth()
-  },
-
-  watch: {
-    value(value) {
-      this.selected = value;
-
-      this.$nextTick(() => {
-        this.calculateWidth()
-      })
-    }
   },
 
   computed: {
@@ -112,14 +120,16 @@ export default {
         "is-large": this.large,
         "is-expanded": this.expanded,
         "is-transparent": this.transparent,
-        "has-border": this.$attrs.hasOwnProperty("border"),
-        level: this.label
+        "has-border": this.border,
+        "level": this.label,
       };
     }
   },
 
   methods: {
     calculateWidth() {
+      if (!this.autoWidth) return
+
       const select = this.$refs.select
 
       if (select.selectedIndex >= 0) {
@@ -133,16 +143,12 @@ export default {
       })
     },
     calculateMaxWidth(width) {
-      if (!this.maxWidth) {
+      if (!this.autoWidthMax) {
         return width
       }
 
-      return width > this.maxWidth ? this.maxWidth : width
+      return width > this.autoWidthMax ? this.autoWidthMax : width
     }
   }
 };
 </script>
-
-<style scoped lang="scss">
-@import "../../assets/sass/components/select.scss";
-</style>
